Updating MARC file and the scripts for updating it

+++ bin/fetch_marc.sh
@@ -4,6 +4,11 @@
 # Fetch MARC records from Voyager and merge them into a single file. All bib
 # ids from ./bib_ids.txt will be retrieved.
 #
+# NOTE: the list of bib ids in bib_ids.txt should be manually updated:
+# * bib ids for fiche items can be found in the shotlist Google doc:
+#     https://docs.google.com/spreadsheets/d/1ffmSwOwKLhxzsdvtJ7mTHFjJshhmCLXQDM6kQCknWdc/edit#gid=0)
+# * bib ids for PUL items can be found in Plum
+#
 # Requires: curl, xmllint
 #
 
@@ -20,4 +25,6 @@ for bib_id in `cat $IDS_FILE`; do
 done
 echo "</collection>" >> $TMP
 xmllint --format --nsclean $TMP > $OUT
-rm $TMP
+if [ $? = 0 ]; then
+  rm $TMP
+fi
